1.0 alpha2-10

+++ app/src/main/java/com/hellguy39/hellweather/retrofit/RetrofitServices.kt
@@ -9,8 +9,8 @@ interface RetrofitServices {
 
     @GET("onecall?")
     fun getWeatherOneCall(
-        @Query("lat")lat: Int,
-        @Query("lon")lon: Int,
+        @Query("lat")lat: Double,
+        @Query("lon")lon: Double,
         @Query("exclude")exclude: String,
         @Query("units")units: String,
         @Query("appid")appId: String
